fix: added comments, cleaned up code
